backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
azure-identity>=1.15.0
openai>=1.12.0
pydantic>=2.6.4
requests>=2.31.0
